+++ packages/pangea-sdk/src/main/java/cloud/pangeacyber/pangea/prompt_guard/requests/GuardRequest.java
@@ -33,9 +33,4 @@ public final class GuardRequest extends BaseRequest {
 
 	/** Boolean to enable classification of the content */
 	boolean classify;
-
-	/**
-	 * Threshold for the confidence score to consider the prompt as malicious
-	 */
-	double threshold;
 }