+++ CHANGELOG.md
@@ -9,7 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
-- Prompt Guard: `classify` and `threshold` parameters.
+- Prompt Guard: `classify` parameter.
 - AI Guard: support for structured input.
 - AI Guard: more detectors.